remove some unused class
