+++ seatunnel-core/seatunnel-flink-starter/src/main/java/org/apache/seatunnel/core/starter/args/FlinkCommandArgs.java
@@ -19,7 +19,6 @@ package org.apache.seatunnel.core.starter.args;
 
 import org.apache.seatunnel.common.config.DeployMode;
 import org.apache.seatunnel.core.starter.command.AbstractCommandArgs;
-import org.apache.seatunnel.core.starter.config.ApiType;
 import org.apache.seatunnel.core.starter.config.EngineType;
 import org.apache.seatunnel.core.starter.config.FlinkRunMode;
 
@@ -34,12 +33,6 @@ public class FlinkCommandArgs extends AbstractCommandArgs {
         converter = RunModeConverter.class,
         description = "job run mode, run or run-application")
     private FlinkRunMode runMode = FlinkRunMode.RUN;
-
-    @Parameter(names = {"-api", "--api-type"},
-        converter = ApiTypeConverter.class,
-        description = "Api type, engine or seatunnel")
-    private ApiType apiType = ApiType.ENGINE_API;
-
     /**
      * Undefined parameters parsed will be stored here as flink command parameters.
      */
@@ -71,14 +64,6 @@ public class FlinkCommandArgs extends AbstractCommandArgs {
         this.flinkParams = flinkParams;
     }
 
-    public ApiType getApiType() {
-        return apiType;
-    }
-
-    public void setApiType(ApiType apiType) {
-        this.apiType = apiType;
-    }
-
     /**
      * Used to convert the run mode string to the enum value.
      */
@@ -97,28 +82,6 @@ public class FlinkCommandArgs extends AbstractCommandArgs {
                 }
             }
             throw new IllegalArgumentException(String.format("Run mode %s not supported", value));
-        }
-    }
-
-    /**
-     * Used to convert the api type string to the enum value.
-     */
-    private static class ApiTypeConverter implements IStringConverter<ApiType> {
-
-        /**
-         * If the '-api' is not set, then will not go into this convert method.
-         *
-         * @param value input value set by '-api' or '--api-type'
-         * @return api type enum value
-         */
-        @Override
-        public ApiType convert(String value) {
-            for (ApiType apiType : ApiType.values()) {
-                if (apiType.getApiType().equalsIgnoreCase(value)) {
-                    return apiType;
-                }
-            }
-            throw new IllegalArgumentException(String.format("API type %s not supported", value));
         }
     }
