+++ seatunnel-core/seatunnel-flink-starter/src/main/java/org/apache/seatunnel/core/starter/command/FlinkCommandBuilder.java
@@ -24,39 +24,21 @@ public class FlinkCommandBuilder implements CommandBuilder<FlinkCommandArgs> {
 
     @Override
     public Command<FlinkCommandArgs> buildCommand(FlinkCommandArgs commandArgs) {
-        if (Boolean.FALSE.equals(Common.setDeployMode(commandArgs.getDeployMode().getName()))) {
+        if (!Common.setDeployMode(commandArgs.getDeployMode().getName())) {
             throw new IllegalArgumentException(
-                String.format("Deploy mode: %s is Illegal", commandArgs.getDeployMode()));
+                    String.format("Deploy mode: %s is Illegal", commandArgs.getDeployMode()));
         }
-        switch (commandArgs.getApiType()) {
-            case ENGINE_API:
-                return new FlinkApiCommandBuilder().buildCommand(commandArgs);
-            case SEATUNNEL_API:
-                return new SeaTunnelApiCommandBuilder().buildCommand(commandArgs);
-            default:
-                throw new IllegalArgumentException("Unsupported API type: " + commandArgs.getApiType());
-        }
-    }
-
-    /**
-     * Used to generate command for engine API.
-     */
-    private static class FlinkApiCommandBuilder extends FlinkCommandBuilder {
-        @Override
-        public Command<FlinkCommandArgs> buildCommand(FlinkCommandArgs commandArgs) {
-            return commandArgs.isCheckConfig() ? new FlinkApiConfValidateCommand(commandArgs)
-                : new SeaTunnelApiTaskExecuteCommand(commandArgs);
-        }
+        return new FlinkApiCommandBuilder().buildCommand(commandArgs);
     }
 
     /**
      * Used to generate command for seaTunnel API.
      */
-    private static class SeaTunnelApiCommandBuilder extends FlinkCommandBuilder {
+    private static class FlinkApiCommandBuilder extends FlinkCommandBuilder {
         @Override
         public Command<FlinkCommandArgs> buildCommand(FlinkCommandArgs commandArgs) {
-            return commandArgs.isCheckConfig() ? new SeaTunnelApiConfValidateCommand(commandArgs)
-                : new SeaTunnelApiTaskExecuteCommand(commandArgs);
+            return commandArgs.isCheckConfig() ? new FlinkApiConfValidateCommand(commandArgs)
+                    : new FlinkApiTaskExecuteCommand(commandArgs);
         }
     }
 }
