+++ seatunnel-examples/seatunnel-flink-examples/src/main/java/org/apache/seatunnel/example/flink/SeaTunnelApiExample.java
@@ -21,7 +21,6 @@ import org.apache.seatunnel.core.starter.Seatunnel;
 import org.apache.seatunnel.core.starter.args.FlinkCommandArgs;
 import org.apache.seatunnel.core.starter.command.Command;
 import org.apache.seatunnel.core.starter.command.FlinkCommandBuilder;
-import org.apache.seatunnel.core.starter.config.ApiType;
 import org.apache.seatunnel.core.starter.exception.CommandException;
 
 import java.io.FileNotFoundException;
@@ -37,14 +36,13 @@ public class SeaTunnelApiExample {
         flinkCommandArgs.setConfigFile(configFile);
         flinkCommandArgs.setCheckConfig(false);
         flinkCommandArgs.setVariables(null);
-        flinkCommandArgs.setApiType(ApiType.SEATUNNEL_API);
         Command<FlinkCommandArgs> flinkCommand =
             new FlinkCommandBuilder().buildCommand(flinkCommandArgs);
         Seatunnel.run(flinkCommand);
     }
 
     public static String getTestConfigFile(String configFile) throws FileNotFoundException, URISyntaxException {
-        URL resource = LocalFlinkExample.class.getResource(configFile);
+        URL resource = SeaTunnelApiExample.class.getResource(configFile);
         if (resource == null) {
             throw new FileNotFoundException("Can't find config file: " + configFile);
         }
